nydus-image: move common inode/chunk wrapper into rafs
Move common inode/chunk wrapper into rafs for reuse.
