+++ src/bin/nydus-image/core/node.rs
@@ -20,26 +20,17 @@ use std::str::FromStr;
 use anyhow::{Context, Error, Result};
 use sha2::digest::Digest;
 
-use nydus_rafs::metadata::cached_v5::{CachedChunkInfoV5, CachedInodeV5};
-use nydus_rafs::metadata::direct_v5::{
-    DirectChunkInfoV5, OndiskInodeWrapper as OndiskInodeWrapperV5,
-};
-use nydus_rafs::metadata::direct_v6::{
-    DirectChunkInfoV6, OndiskInodeWrapper as OndiskInodeWrapperV6,
-};
-use nydus_rafs::metadata::layout::v5::{
-    RafsV5ChunkInfo, RafsV5Inode, RafsV5InodeFlags, RafsV5InodeWrapper,
-};
+use nydus_rafs::metadata::chunk::ChunkWrapper;
+use nydus_rafs::metadata::inode::InodeWrapper;
+use nydus_rafs::metadata::layout::v5::RafsV5InodeWrapper;
 use nydus_rafs::metadata::layout::v6::{
     align_offset, calculate_nid, RafsV6Dirent, RafsV6InodeChunkAddr, RafsV6InodeChunkHeader,
     RafsV6InodeCompact, RafsV6InodeExtended, RafsV6OndiskInode, EROFS_BLOCK_SIZE,
     EROFS_INODE_CHUNK_BASED, EROFS_INODE_FLAT_INLINE, EROFS_INODE_FLAT_PLAIN,
 };
 use nydus_rafs::metadata::layout::RafsXAttrs;
-use nydus_rafs::metadata::{Inode, RafsInodeExt, RafsStore};
+use nydus_rafs::metadata::{Inode, RafsStore, RafsVersion};
 use nydus_rafs::RafsIoWrite;
-use nydus_storage::device::v5::BlobV5ChunkInfo;
-use nydus_storage::device::{BlobChunkFlags, BlobChunkInfo};
 use nydus_utils::{
     compress,
     digest::{DigestHasher, RafsDigest},
@@ -47,9 +38,7 @@ use nydus_utils::{
 };
 
 use super::chunk_dict::{ChunkDict, DigestWithBlobIndex};
-use super::context::{
-    ArtifactWriter, BlobContext, BlobManager, BootstrapContext, BuildContext, RafsVersion,
-};
+use super::context::{ArtifactWriter, BlobContext, BlobManager, BootstrapContext, BuildContext};
 use super::tree::Tree;
 
 // Filesystem may have different algorithms to calculate `i_size` for directory entries,
@@ -1365,620 +1354,6 @@ impl Node {
         }
 
         None
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum InodeWrapper {
-    V5(RafsV5Inode),
-    // Reuse `RafsV5Inode` for v6 with a different wrapper to reduce duplicated code.
-    V6(RafsV5Inode),
-}
-
-impl InodeWrapper {
-    pub fn new(version: RafsVersion) -> Self {
-        match version {
-            RafsVersion::V5 => InodeWrapper::V5(RafsV5Inode::new()),
-            RafsVersion::V6 => InodeWrapper::V6(RafsV5Inode::new()),
-        }
-    }
-
-    pub fn from_inode_info(inode: &dyn RafsInodeExt) -> Self {
-        if let Some(inode) = inode.as_any().downcast_ref::<CachedInodeV5>() {
-            InodeWrapper::V5(to_rafsv5_inode(inode))
-        } else if let Some(inode) = inode.as_any().downcast_ref::<OndiskInodeWrapperV5>() {
-            InodeWrapper::V5(to_rafsv5_inode(inode))
-        } else if let Some(inode) = inode.as_any().downcast_ref::<OndiskInodeWrapperV6>() {
-            InodeWrapper::V6(to_rafsv5_inode(inode))
-        } else {
-            panic!("unknown inode information struct");
-        }
-    }
-
-    pub fn inode_size(&self) -> usize {
-        match self {
-            InodeWrapper::V5(i) => i.size(),
-            InodeWrapper::V6(i) => i.size(),
-        }
-    }
-
-    pub fn mode(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.mode(),
-            InodeWrapper::V6(i) => i.mode(),
-        }
-    }
-
-    pub fn set_mode(&mut self, mode: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_mode = mode,
-            InodeWrapper::V6(i) => i.i_mode = mode,
-        }
-    }
-
-    pub fn is_dir(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.is_dir(),
-            InodeWrapper::V6(i) => i.is_dir(),
-        }
-    }
-
-    pub fn is_reg(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.is_reg(),
-            InodeWrapper::V6(i) => i.is_reg(),
-        }
-    }
-
-    pub fn is_hardlink(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.is_hardlink(),
-            InodeWrapper::V6(i) => i.is_hardlink(),
-        }
-    }
-
-    pub fn is_symlink(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.is_symlink(),
-            InodeWrapper::V6(i) => i.is_symlink(),
-        }
-    }
-
-    pub fn is_chrdev(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFCHR as u32,
-            InodeWrapper::V6(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFCHR as u32,
-        }
-    }
-
-    pub fn is_blkdev(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFBLK as u32,
-            InodeWrapper::V6(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFBLK as u32,
-        }
-    }
-
-    pub fn is_fifo(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFIFO as u32,
-            InodeWrapper::V6(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFIFO as u32,
-        }
-    }
-
-    pub fn is_sock(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFSOCK as u32,
-            InodeWrapper::V6(i) => i.i_mode & libc::S_IFMT as u32 == libc::S_IFSOCK as u32,
-        }
-    }
-
-    pub fn is_special(&self) -> bool {
-        self.is_chrdev() || self.is_blkdev() || self.is_fifo() || self.is_sock()
-    }
-
-    pub fn has_xattr(&self) -> bool {
-        match self {
-            InodeWrapper::V5(i) => i.has_xattr(),
-            InodeWrapper::V6(i) => i.has_xattr(),
-        }
-    }
-
-    pub fn set_has_xattr(&mut self, enable: bool) {
-        match self {
-            InodeWrapper::V5(i) => {
-                if enable {
-                    i.i_flags |= RafsV5InodeFlags::XATTR;
-                } else {
-                    i.i_flags &= !RafsV5InodeFlags::XATTR;
-                }
-            }
-            InodeWrapper::V6(i) => {
-                if enable {
-                    i.i_flags |= RafsV5InodeFlags::XATTR;
-                } else {
-                    i.i_flags &= !RafsV5InodeFlags::XATTR;
-                }
-            }
-        }
-    }
-
-    pub fn ino(&self) -> Inode {
-        match self {
-            InodeWrapper::V5(i) => i.i_ino,
-            InodeWrapper::V6(i) => i.i_ino,
-        }
-    }
-
-    pub fn set_ino(&mut self, ino: Inode) {
-        match self {
-            InodeWrapper::V5(i) => i.i_ino = ino,
-            InodeWrapper::V6(i) => i.i_ino = ino,
-        }
-    }
-
-    pub fn parent(&self) -> Inode {
-        match self {
-            InodeWrapper::V5(i) => i.i_parent,
-            InodeWrapper::V6(i) => i.i_parent,
-        }
-    }
-
-    pub fn set_parent(&mut self, parent: Inode) {
-        match self {
-            InodeWrapper::V5(i) => i.i_parent = parent,
-            InodeWrapper::V6(i) => i.i_parent = parent,
-        }
-    }
-
-    pub fn size(&self) -> u64 {
-        match self {
-            InodeWrapper::V5(i) => i.i_size,
-            InodeWrapper::V6(i) => i.i_size,
-        }
-    }
-
-    pub fn set_size(&mut self, size: u64) {
-        match self {
-            InodeWrapper::V5(i) => i.i_size = size,
-            InodeWrapper::V6(i) => i.i_size = size,
-        }
-    }
-
-    pub fn uid(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_uid,
-            InodeWrapper::V6(i) => i.i_uid,
-        }
-    }
-
-    pub fn set_uid(&mut self, uid: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_uid = uid,
-            InodeWrapper::V6(i) => i.i_uid = uid,
-        }
-    }
-
-    pub fn gid(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_gid,
-            InodeWrapper::V6(i) => i.i_gid,
-        }
-    }
-
-    pub fn set_gid(&mut self, gid: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_gid = gid,
-            InodeWrapper::V6(i) => i.i_gid = gid,
-        }
-    }
-
-    pub fn mtime(&self) -> u64 {
-        match self {
-            InodeWrapper::V5(i) => i.i_mtime,
-            InodeWrapper::V6(i) => i.i_mtime,
-        }
-    }
-
-    pub fn set_mtime(&mut self, mtime: u64) {
-        match self {
-            InodeWrapper::V5(i) => i.i_mtime = mtime,
-            InodeWrapper::V6(i) => i.i_mtime = mtime,
-        }
-    }
-
-    pub fn mtime_nsec(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_mtime_nsec,
-            InodeWrapper::V6(i) => i.i_mtime_nsec,
-        }
-    }
-
-    pub fn set_mtime_nsec(&mut self, mtime_nsec: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_mtime_nsec = mtime_nsec,
-            InodeWrapper::V6(i) => i.i_mtime_nsec = mtime_nsec,
-        }
-    }
-
-    pub fn blocks(&self) -> u64 {
-        match self {
-            InodeWrapper::V5(i) => i.i_blocks,
-            InodeWrapper::V6(i) => i.i_blocks,
-        }
-    }
-
-    pub fn set_blocks(&mut self, blocks: u64) {
-        match self {
-            InodeWrapper::V5(i) => i.i_blocks = blocks,
-            InodeWrapper::V6(i) => i.i_blocks = blocks,
-        }
-    }
-
-    pub fn set_rdev(&mut self, rdev: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_rdev = rdev,
-            InodeWrapper::V6(i) => i.i_rdev = rdev,
-        }
-    }
-
-    pub fn set_projid(&mut self, projid: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_projid = projid,
-            InodeWrapper::V6(i) => i.i_projid = projid,
-        }
-    }
-
-    pub fn nlink(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_nlink,
-            InodeWrapper::V6(i) => i.i_nlink,
-        }
-    }
-
-    pub fn set_nlink(&mut self, nlink: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_nlink = nlink,
-            InodeWrapper::V6(i) => i.i_nlink = nlink,
-        }
-    }
-
-    pub fn digest(&self) -> &RafsDigest {
-        match self {
-            InodeWrapper::V5(i) => &i.i_digest,
-            InodeWrapper::V6(i) => &i.i_digest,
-        }
-    }
-
-    pub fn set_digest(&mut self, digest: RafsDigest) {
-        match self {
-            InodeWrapper::V5(i) => i.i_digest = digest,
-            InodeWrapper::V6(i) => i.i_digest = digest,
-        }
-    }
-
-    pub fn name_size(&self) -> u16 {
-        match self {
-            InodeWrapper::V5(i) => i.i_name_size,
-            InodeWrapper::V6(i) => i.i_name_size,
-        }
-    }
-
-    pub fn set_name_size(&mut self, size: usize) {
-        debug_assert!(size < u16::MAX as usize);
-        match self {
-            InodeWrapper::V5(i) => i.i_name_size = size as u16,
-            InodeWrapper::V6(i) => i.i_name_size = size as u16,
-        }
-    }
-
-    pub fn symlink_size(&self) -> u16 {
-        match self {
-            InodeWrapper::V5(i) => i.i_symlink_size,
-            InodeWrapper::V6(i) => i.i_symlink_size,
-        }
-    }
-
-    pub fn set_symlink_size(&mut self, size: usize) {
-        debug_assert!(size <= u16::MAX as usize);
-        match self {
-            InodeWrapper::V5(i) => {
-                i.i_flags |= RafsV5InodeFlags::SYMLINK;
-                i.i_symlink_size = size as u16;
-            }
-            InodeWrapper::V6(i) => {
-                i.i_flags |= RafsV5InodeFlags::SYMLINK;
-                i.i_symlink_size = size as u16;
-            }
-        }
-    }
-
-    pub fn child_index(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_child_index,
-            InodeWrapper::V6(i) => i.i_child_index,
-        }
-    }
-
-    pub fn set_child_index(&mut self, index: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_child_index = index,
-            InodeWrapper::V6(i) => i.i_child_index = index,
-        }
-    }
-
-    pub fn child_count(&self) -> u32 {
-        match self {
-            InodeWrapper::V5(i) => i.i_child_count,
-            InodeWrapper::V6(i) => i.i_child_count,
-        }
-    }
-
-    pub fn set_child_count(&mut self, count: u32) {
-        match self {
-            InodeWrapper::V5(i) => i.i_child_count = count,
-            InodeWrapper::V6(i) => i.i_child_count = count,
-        }
-    }
-
-    fn create_chunk(&self) -> ChunkWrapper {
-        match self {
-            InodeWrapper::V5(_) => ChunkWrapper::V5(RafsV5ChunkInfo::new()),
-            InodeWrapper::V6(_) => ChunkWrapper::V6(RafsV5ChunkInfo::new()),
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum ChunkWrapper {
-    V5(RafsV5ChunkInfo),
-    // Reuse `RafsV5ChunkInfo` for v6 with a different wrapper to reduce duplicated code.
-    V6(RafsV5ChunkInfo),
-}
-
-impl ChunkWrapper {
-    #[cfg(test)]
-    pub fn new(version: RafsVersion) -> Self {
-        match version {
-            RafsVersion::V5 => ChunkWrapper::V5(RafsV5ChunkInfo::default()),
-            RafsVersion::V6 => ChunkWrapper::V6(RafsV5ChunkInfo::default()),
-        }
-    }
-
-    pub fn from_chunk_info(cki: &dyn BlobChunkInfo) -> Self {
-        if let Some(cki_v5) = cki.as_any().downcast_ref::<CachedChunkInfoV5>() {
-            ChunkWrapper::V5(to_rafsv5_chunk_info(cki_v5))
-        } else if let Some(cki_v5) = cki.as_any().downcast_ref::<DirectChunkInfoV5>() {
-            ChunkWrapper::V5(to_rafsv5_chunk_info(cki_v5))
-        } else if let Some(cki_v6) = cki.as_any().downcast_ref::<DirectChunkInfoV6>() {
-            ChunkWrapper::V6(to_rafsv5_chunk_info(cki_v6))
-        } else {
-            panic!("unknown chunk information struct");
-        }
-    }
-
-    pub fn id(&self) -> &RafsDigest {
-        match self {
-            ChunkWrapper::V5(c) => &c.block_id,
-            ChunkWrapper::V6(c) => &c.block_id,
-        }
-    }
-
-    pub fn set_id(&mut self, id: RafsDigest) {
-        match self {
-            ChunkWrapper::V5(c) => c.block_id = id,
-            ChunkWrapper::V6(c) => c.block_id = id,
-        }
-    }
-
-    pub fn index(&self) -> u32 {
-        match self {
-            ChunkWrapper::V5(c) => c.index,
-            ChunkWrapper::V6(c) => c.index,
-        }
-    }
-
-    pub fn set_index(&mut self, index: u32) {
-        match self {
-            ChunkWrapper::V5(c) => c.index = index,
-            ChunkWrapper::V6(c) => c.index = index,
-        }
-    }
-
-    pub fn blob_index(&self) -> u32 {
-        match self {
-            ChunkWrapper::V5(c) => c.blob_index,
-            ChunkWrapper::V6(c) => c.blob_index,
-        }
-    }
-
-    pub fn set_blob_index(&mut self, index: u32) {
-        match self {
-            ChunkWrapper::V5(c) => c.blob_index = index,
-            ChunkWrapper::V6(c) => c.blob_index = index,
-        }
-    }
-
-    pub fn compressed_offset(&self) -> u64 {
-        match self {
-            ChunkWrapper::V5(c) => c.compressed_offset,
-            ChunkWrapper::V6(c) => c.compressed_offset,
-        }
-    }
-
-    pub fn set_compressed_offset(&mut self, offset: u64) {
-        match self {
-            ChunkWrapper::V5(c) => c.compressed_offset = offset,
-            ChunkWrapper::V6(c) => c.compressed_offset = offset,
-        }
-    }
-
-    pub fn compressed_size(&self) -> u32 {
-        match self {
-            ChunkWrapper::V5(c) => c.compressed_size,
-            ChunkWrapper::V6(c) => c.compressed_size,
-        }
-    }
-
-    pub fn uncompressed_offset(&self) -> u64 {
-        match self {
-            ChunkWrapper::V5(c) => c.uncompressed_offset,
-            ChunkWrapper::V6(c) => c.uncompressed_offset,
-        }
-    }
-
-    pub fn set_uncompressed_offset(&mut self, offset: u64) {
-        match self {
-            ChunkWrapper::V5(c) => c.uncompressed_offset = offset,
-            ChunkWrapper::V6(c) => c.uncompressed_offset = offset,
-        }
-    }
-
-    pub fn uncompressed_size(&self) -> u32 {
-        match self {
-            ChunkWrapper::V5(c) => c.uncompressed_size,
-            ChunkWrapper::V6(c) => c.uncompressed_size,
-        }
-    }
-
-    pub fn file_offset(&self) -> u64 {
-        match self {
-            ChunkWrapper::V5(c) => c.file_offset,
-            ChunkWrapper::V6(c) => c.file_offset,
-        }
-    }
-
-    pub fn set_file_offset(&mut self, offset: u64) {
-        match self {
-            ChunkWrapper::V5(c) => c.file_offset = offset,
-            ChunkWrapper::V6(c) => c.file_offset = offset,
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    #[inline]
-    fn set_chunk_info(
-        &mut self,
-        blob_index: u32,
-        chunk_index: u32,
-        file_offset: u64,
-        uncompressed_offset: u64,
-        uncompressed_size: u32,
-        compressed_offset: u64,
-        compressed_size: u32,
-        is_compressed: bool,
-    ) -> Result<()> {
-        match self {
-            ChunkWrapper::V5(c) => {
-                c.index = chunk_index;
-                c.blob_index = blob_index;
-                c.file_offset = file_offset;
-                c.compressed_offset = compressed_offset;
-                c.compressed_size = compressed_size;
-                c.uncompressed_offset = uncompressed_offset;
-                c.uncompressed_size = uncompressed_size;
-                if is_compressed {
-                    c.flags |= BlobChunkFlags::COMPRESSED;
-                }
-            }
-            ChunkWrapper::V6(c) => {
-                c.index = chunk_index;
-                c.blob_index = blob_index;
-                c.file_offset = file_offset;
-                c.compressed_offset = compressed_offset;
-                c.compressed_size = compressed_size;
-                c.uncompressed_offset = uncompressed_offset;
-                c.uncompressed_size = uncompressed_size;
-                if is_compressed {
-                    c.flags |= BlobChunkFlags::COMPRESSED;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn copy_from(&mut self, other: &Self) {
-        match (self, other) {
-            (ChunkWrapper::V5(s), ChunkWrapper::V5(o)) => {
-                s.clone_from(o);
-            }
-            (ChunkWrapper::V6(s), ChunkWrapper::V6(o)) => {
-                s.clone_from(o);
-            }
-            (ChunkWrapper::V5(s), ChunkWrapper::V6(o)) => {
-                s.clone_from(o);
-            }
-            (ChunkWrapper::V6(s), ChunkWrapper::V5(o)) => {
-                s.clone_from(o);
-            }
-        }
-    }
-
-    pub(crate) fn store(&self, w: &mut dyn RafsIoWrite) -> Result<usize> {
-        match self {
-            ChunkWrapper::V5(c) => c.store(w).context("failed to store rafs v5 chunk"),
-            ChunkWrapper::V6(c) => c.store(w).context("failed to store rafs v6 chunk"),
-        }
-    }
-}
-
-impl Display for ChunkWrapper {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "id {}, index {}, blob_index {}, file_offset {}, compressed {}/{}, uncompressed {}/{}",
-            self.id(),
-            self.index(),
-            self.blob_index(),
-            self.file_offset(),
-            self.compressed_offset(),
-            self.compressed_size(),
-            self.uncompressed_offset(),
-            self.uncompressed_size(),
-        )
-    }
-}
-
-/// Construct a `RafsV5Inode` object from a `Arc<dyn RafsInode>` object.
-fn to_rafsv5_inode(inode: &dyn RafsInodeExt) -> RafsV5Inode {
-    let attr = inode.get_attr();
-
-    RafsV5Inode {
-        i_digest: inode.get_digest(),
-        i_parent: inode.parent(),
-        i_ino: attr.ino,
-        i_uid: attr.uid,
-        i_gid: attr.gid,
-        i_projid: inode.projid(),
-        i_mode: attr.mode,
-        i_size: attr.size,
-        i_blocks: attr.blocks,
-        i_flags: RafsV5InodeFlags::from_bits_truncate(inode.flags()),
-        i_nlink: attr.nlink,
-        i_child_index: inode.get_child_index().unwrap_or(0),
-        i_child_count: inode.get_child_count(),
-        i_name_size: inode.get_name_size(),
-        i_symlink_size: inode.get_symlink_size(),
-        i_rdev: attr.rdev,
-        i_mtime_nsec: attr.mtimensec,
-        i_mtime: attr.mtime,
-        i_reserved: [0u8; 8],
-    }
-}
-
-/// Construct a `RafsV5ChunkInfo` object from a `dyn BlobChunkInfo` object.
-fn to_rafsv5_chunk_info(cki: &dyn BlobV5ChunkInfo) -> RafsV5ChunkInfo {
-    RafsV5ChunkInfo {
-        block_id: *cki.chunk_id(),
-        blob_index: cki.blob_index(),
-        flags: cki.flags(),
-        compressed_size: cki.compressed_size(),
-        uncompressed_size: cki.uncompressed_size(),
-        compressed_offset: cki.compressed_offset(),
-        uncompressed_offset: cki.uncompressed_offset(),
-        file_offset: cki.file_offset(),
-        index: cki.index(),
-        reserved: 0u32,
     }
 }
